+++ client/src/helper/urls.js
@@ -113,7 +113,6 @@ const urls = {
   },
   events: (regionId) => `/region?bid=${regionId}&sub=events`,
   addEvents: (regionId) => `/?page=event&sub=add&bid=${regionId}`,
-  foodsaverList: (regionId) => `/?page=foodsaver&bid=${regionId}`,
   foodsharepoints: (regionId) => `/region?bid=${regionId}&sub=fairteiler`,
   members: (regionId) => `/region?bid=${regionId}&sub=members`,
   options: (regionId) => `/region?bid=${regionId}&sub=options`,