+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - REST endpoint for editing content entries !3098 @alex.simm
 - Created confirmationDialogue mixin !3140 @AntonBallmaier
 - Removed unused event invitation code #1739 !3149 @AntonBallmaier
+- Removed duplicate region member page and profile XHR !3169 @AntonBallmaier
 - Move Workgroup is Admin Check method from session to module #1769 !3168 @KoeMai
 - Move location information in session to user #1769 !3148 @KoeMai
 - Remove unused variables from session #1769 !3161  @KoeMai